HistoryModelUpdate

+++ Models/Room.cs
@@ -11,9 +11,9 @@ namespace GypooWebAPI.Models
         public string roomType { get; set; } = null!;
         public int guest { get; set; } = 0!;
         public int roomCount { get; set; } = 0!;
-        public int currentRoom { get; set; } = 0!;
+        public int currentRoom { get; set; } = 0!; // เอามาจาก roomAva
         public List<string> picture { get; set; } = new List<string>();
-        public float roomPrice { get; set; } = 0!;
+        public double roomPrice { get; set; } = 0!;
         public int discount { get; set; } = 0!;
         public List<RoomAva> roomCount30Day { get; set; } = new List<RoomAva>();
         public ServiceRoom service { get; set; } = null!;
